requirements.txt
click==7.1.2
pika==1.1.0
python-dotenv==0.14.0
tqdm==4.48.2
zstandard==0.14.0
